agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.3.7
anthropic==0.8.1
google-generativeai==0.3.2
httpx==0.25.2
python-multipart==0.0.6
python-json-logger==2.0.7

